==> 육아 정보 공유 플랫폼_조타/0. code/5. MySQL insert.ipynb <==
import pandas as pd

data = pd.read_csv('./2. deeplearning/분유 제조사별 선호도.csv', encoding = 'utf-8')

import mysql.connector
from sqlalchemy import create_engine
import sqlite3

engine = create_engine("mysql+pymysql://root:"+""+"@localhost/jota"+"?charset=utf8")
conn = engine.connect()
data.to_sql(name='milk', if_exists='replace', con=engine, index=True)

import pandas as pd

data = pd.read_csv('./2. deeplearning/이유식 제조사별 선호도.csv', encoding = 'utf-8')

import mysql.connector
from sqlalchemy import create_engine
import sqlite3

engine = create_engine("mysql+pymysql://root:"+""+"@localhost/jota"+"?charset=utf8")
conn = engine.connect()
data.to_sql(name='slowfood', if_exists='replace', con=engine, index=True)

import pandas as pd

data = pd.read_csv('./2. deeplearning/기저귀 제조사별 선호도.csv', encoding = 'utf-8')

import mysql.connector
from sqlalchemy import create_engine
import sqlite3

engine = create_engine("mysql+pymysql://root:"+""+"@localhost/jota"+"?charset=utf8")
conn = engine.connect()
data.to_sql(name='gijugi', if_exists='replace', con=engine, index=True)

import pandas as pd

data = pd.read_csv('./2. deeplearning/유모차 제조사별 선호도.csv', encoding = 'utf-8')

import mysql.connector
from sqlalchemy import create_engine
import sqlite3

engine = create_engine("mysql+pymysql://root:"+""+"@localhost/jota"+"?charset=utf8")
conn = engine.connect()
data.to_sql(name='babycar', if_exists='replace', con=engine, index=True)
